backend/requirements.txt
Django==5.0.2
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
requests==2.31.0
python-dotenv==1.0.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.5.0